requirements.txt
SpeechRecognition==3.10.3
moviepy==1.0.3
ffmpeg-python==0.2.0